Split diegest out of sha, since it will be used for other hashes
