+++ risc0/zkp/src/core/mod.rs
@@ -24,6 +24,7 @@
 
 extern crate alloc;
 
+pub mod digest;
 pub mod ntt;
 pub mod poly;
 pub mod sha;
